chore: re-audit data types

+++ packages/core/src/model/trading/position.rs
@@ -7,6 +7,6 @@ pub struct Position {
     pub symbol: Symbol,
     pub currency: Currency,
     pub cost_price: Decimal,
-    pub total_quantity: i64,
-    pub available_quantity: i64,
+    pub total_quantity: Decimal,
+    pub available_quantity: Decimal,
 }
